[test] lightning trainer to use the new datamodule api
[ghstack-poisoned]
