+++ requirements.txt
@@ -15,7 +15,7 @@ omegaconf==2.0.6
 lmdb==0.98
 termcolor==1.1.0
 iopath==0.1.7
-pytorch_lightning==1.1.6
+pytorch_lightning==1.2.7
 datasets==1.2.1
 matplotlib==3.3.4
 pycocotools==2.0.2
