+++ tests/trainers/lightning/test_grad_clipping.py
@@ -6,7 +6,7 @@ from unittest.mock import MagicMock
 import torch
 from mmf.utils.general import clip_gradients
 from pytorch_lightning.callbacks.base import Callback
-from tests.trainers.lightning.test_utils import get_lightning_trainer, get_mmf_trainer
+from tests.trainers.test_utils import get_lightning_trainer, get_mmf_trainer
 
 
 class TestLightningTrainerGradClipping(unittest.TestCase, Callback):
